[IDLE-122] refactor: 패키지 구조 변경으로 인해 파일 이동
