+++ backend/book-service/src/main/java/kr/mybrary/bookservice/global/dto/response/SuccessResponse.java
@@ -1,4 +1,4 @@
-package kr.mybrary.bookservice.global;
+package kr.mybrary.bookservice.global.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
